+++ robot_descriptions/go1_description.py
@@ -16,7 +16,7 @@
 # limitations under the License.
 
 """
-Aliengo MJCF description.
+Go1 MJCF description.
 """
 
 from os import path as _path
@@ -25,8 +25,10 @@ from ._cache import clone_to_cache as _clone_to_cache
 
 REPOSITORY_PATH: str = _clone_to_cache("unitree_mujoco")
 
-PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "data", "aliengo")
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "data", "go1")
 
 MESHES_PATH: str = _path.join(PACKAGE_PATH, "meshes")
 
-MJCF_PATH: str = _path.join(PACKAGE_PATH, "xml", "aliengo.xml")
+MJCF_PATH: str = _path.join(PACKAGE_PATH, "xml", "go1.xml")
+
+URDF_PATH: str = _path.join(PACKAGE_PATH, "urdf", "go1.urdf")