+++ robot_descriptions/_description_names.py
@@ -61,4 +61,5 @@ DESCRIPTION_NAMES: List[str] = [
     "ur10_description",
     "ur3_description",
     "ur5_description",
+    "ur5e_mj_description",
 ]